feat: laps Float 데이터타입 수정 및 CustomValidator (#116)

+++ module-domain/src/main/java/com/depromeet/memory/port/in/command/CreateStrokeCommand.java
@@ -1,3 +1,3 @@
 package com.depromeet.memory.port.in.command;
 
-public record CreateStrokeCommand(String name, Short laps, Integer meter) {}
+public record CreateStrokeCommand(String name, Float laps, Integer meter) {}
