+++ module-domain/src/main/java/com/depromeet/memory/port/in/command/UpdateStrokeCommand.java
@@ -1,3 +1,3 @@
 package com.depromeet.memory.port.in.command;
 
-public record UpdateStrokeCommand(Long id, String name, Short laps, Integer meter) {}
+public record UpdateStrokeCommand(Long id, String name, Float laps, Integer meter) {}
